+++ leader/leader.go
@@ -3,6 +3,9 @@ package leader
 import (
 	"context"
 	"fmt"
+	sidecar "github.com/lightstep/opentelemetry-prometheus-sidecar"
+	"github.com/lightstep/opentelemetry-prometheus-sidecar/config"
+	"go.opentelemetry.io/otel/metric"
 	"time"
 
 	"github.com/go-kit/kit/log"
@@ -17,6 +20,7 @@ import (
 
 type Candidate interface {
 	Start(ctx context.Context) error
+	IsLeader() bool
 }
 
 type Controller interface {
@@ -26,15 +30,30 @@ type Controller interface {
 }
 
 type candidate struct {
-	client  kubernetes.Interface
-	ctrl    Controller
-	id      string
-	elector *leaderelection.LeaderElector
-	logger  log.Logger
+	client       kubernetes.Interface
+	ctrl         Controller
+	id           string
+	elector      *leaderelection.LeaderElector
+	logger       log.Logger
+	leaderMetric metric.Int64UpDownSumObserver
 }
 
 type LoggingController struct {
 	log.Logger
+}
+
+func NewAlwaysLeaderCandidate() Candidate {
+	return alwaysLeader{}
+}
+
+type alwaysLeader struct{}
+
+func (a alwaysLeader) Start(_ context.Context) error {
+	return nil
+}
+
+func (a alwaysLeader) IsLeader() bool {
+	return true
 }
 
 func NewClient() (*kubernetes.Clientset, error) {
@@ -49,13 +68,25 @@ func NewClient() (*kubernetes.Clientset, error) {
 	return client, err
 }
 
-func NewCandidate(client kubernetes.Interface, namespace, name, id string, ctrl Controller, logger log.Logger) (Candidate, error) {
+func NewKubernetesCandidate(client kubernetes.Interface, namespace, name, id string, ctrl Controller, logger log.Logger) (Candidate, error) {
 	c := &candidate{
 		client: client,
 		ctrl:   ctrl,
 		id:     id,
 		logger: logger,
 	}
+
+	c.leaderMetric = sidecar.OTelMeterMust.NewInt64UpDownSumObserver(
+		config.LeadershipMetric,
+		func(ctx context.Context, result metric.Int64ObserverResult) {
+			if c.IsLeader() {
+				result.Observe(1)
+			} else {
+				result.Observe(0)
+			}
+		},
+		metric.WithDescription("Leadership status of this sidecar"),
+	)
 
 	lock := &resourcelock.LeaseLock{
 		LeaseMeta: metav1.ObjectMeta{
@@ -76,10 +107,14 @@ func NewCandidate(client kubernetes.Interface, namespace, name, id string, ctrl 
 		RenewDeadline:   15 * time.Second,
 		RetryPeriod:     5 * time.Second,
 		Callbacks: leaderelection.LeaderCallbacks{
-			OnStartedLeading: ctrl.OnStartedLeading,
-			OnStoppedLeading: ctrl.OnStoppedLeading,
+			OnStartedLeading: func(ctx context.Context) {
+				ctrl.OnStartedLeading(ctx)
+			},
+			OnStoppedLeading: func() {
+				ctrl.OnStoppedLeading()
+			},
 			OnNewLeader: func(id string) {
-				ctrl.OnNewLeader(lock.LockConfig.Identity == c.id, id)
+				ctrl.OnNewLeader(lock.LockConfig.Identity == id, id)
 			},
 		},
 	}
@@ -98,6 +133,12 @@ func (c *candidate) Start(ctx context.Context) error {
 	go c.elector.Run(ctx)
 
 	return nil
+}
+
+var _ Candidate = (*alwaysLeader)(nil)
+
+func (c *candidate) IsLeader() bool {
+	return c.elector.IsLeader()
 }
 
 func (c LoggingController) OnStartedLeading(ctx context.Context) {